requirements.txt
iperf3
influxdb-client
pyyaml
schema
speedtest-cli